+++ room-hub/server.go
@@ -1,73 +1,21 @@
 package main
 
 import (
-	"context"
-	"crypto/rand"
 	"encoding/json"
-	"errors"
+	"github.com/go-playground/validator/v10"
+	"github.com/gorilla/websocket"
 	"log"
-	"math/big"
 	"net/http"
 	"os"
-	"sync"
-	"time"
-
-	"github.com/go-playground/validator/v10"
-
-	"github.com/google/uuid"
-	"github.com/gorilla/websocket"
 )
 
-const pinRotationInterval = 10 // Interval in seconds
-
 var validate *validator.Validate
-
-const Digits string = "0123456789"
-const ASCIILettersUppercase string = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
-
-// generate a random string of given length from a given character set
-func randomString(length int, characterSet string) (string, error) {
-	var generatedString = ""
-	for j := 0; j < length; j++ {
-		n, err := rand.Int(rand.Reader, big.NewInt(int64(len(characterSet))))
-		if err != nil {
-			log.Printf("failed to generate random string: %s", err)
-			return "", err
-		}
-		generatedString += string(characterSet[n.Int64()])
-	}
-	return generatedString, nil
-}
 
 // Upgrader to upgrade HTTP connections to WebSocket
 var upgrader = websocket.Upgrader{
 	ReadBufferSize:  1024,
 	WriteBufferSize: 1024,
 	CheckOrigin:     func(r *http.Request) bool { return true },
-}
-
-// Struct to hold connection pairs and mutex for thread safety
-type ConnectionManager struct {
-	sync.Mutex
-	rooms     map[string]*Room
-	plugins   map[string]*Plugin
-	pinToRoom map[string]string // Used to look up roomID via room PIN
-}
-
-// Struct to store room connection and configuration
-type Room struct {
-	Id         string
-	Conn       *websocket.Conn
-	PluginID   string
-	Config     RoomConfig
-	State      string
-	StopPinGen context.CancelFunc
-}
-
-type Plugin struct {
-	Id     string
-	RoomID string
-	Conn   *websocket.Conn
 }
 
 // Initialize ConnectionManager instance
@@ -77,85 +25,6 @@ var connManager = &ConnectionManager{
 	pinToRoom: make(map[string]string),
 }
 
-// Function to generate a new room ID
-func generatePIN() (string, error) {
-	const PINLength int = 6
-	const maxTries = 1000
-
-	for try := 0; try < maxTries; try++ {
-
-		var pin, err = randomString(PINLength, Digits)
-		if err != nil {
-			log.Printf("failed to generate room ID: %s", err)
-			return "", err
-		}
-
-		_, PINexists := connManager.pinToRoom[pin]
-		if !PINexists {
-			return pin, nil
-		}
-	}
-	// No pin generated after maxTries
-	return "", errors.New("PIN generation failed")
-}
-
-func generateVerificationCode() (string, error) {
-	const length int = 4
-	return randomString(length, ASCIILettersUppercase+Digits)
-}
-
-func rotatePIN(room *Room, oldPIN string) (newPIN string) {
-	connManager.Lock()
-	// Generate a new PIN
-	newPIN, err := generatePIN()
-	if err != nil {
-		log.Println("Error generating new PIN:", err)
-		connManager.Unlock()
-		return oldPIN
-	}
-	connManager.pinToRoom[newPIN] = room.Id
-	// Remove the old pin from mapping
-	delete(connManager.pinToRoom, oldPIN)
-	connManager.Unlock()
-
-	roomPINMessageJSON, err := createPairingPINMessage(newPIN)
-	if err != nil {
-		room.Conn.Close()
-		return ""
-	}
-
-	if err := room.Conn.WriteMessage(websocket.TextMessage, roomPINMessageJSON); err != nil {
-		log.Println("write error:", err)
-		room.Conn.Close()
-		return ""
-	}
-
-	log.Printf("Rotated PIN for roomId: %s, pin: %s", room.Id, newPIN)
-	return newPIN
-}
-
-// Forward messages between two WebSocket connections
-func processMessageForwarding(src *websocket.Conn, dest *websocket.Conn, done chan struct{}, once *sync.Once) {
-	defer func() {
-		once.Do(func() { close(done) })
-		src.Close()
-		dest.Close()
-	}()
-
-	for {
-		_, msg, err := src.ReadMessage()
-		if err != nil {
-			log.Println("read error:", err)
-			return
-		}
-
-		if err := dest.WriteMessage(websocket.TextMessage, msg); err != nil {
-			log.Println("write error:", err)
-			return
-		}
-	}
-}
-
 func roomHandler(w http.ResponseWriter, r *http.Request) {
 	roomConn, err := upgrader.Upgrade(w, r, nil)
 	if err != nil {
@@ -163,18 +32,9 @@ func roomHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	// Generate room ID
-	var roomID = uuid.New().String()
-
-	room := &Room{
-		Id:    roomID,
-		Conn:  roomConn,
-		State: "new",
-	}
-
-	connManager.Lock()
-	connManager.rooms[roomID] = room
-	connManager.Unlock()
+	// Create a new room
+	room := newRoom(roomConn)
+	connManager.addRoom(room)
 
 	// Handle incoming messages from the room
 	for {
@@ -182,7 +42,7 @@ func roomHandler(w http.ResponseWriter, r *http.Request) {
 		if err != nil {
 			log.Println("read error:", err)
 
-			handleRoomDisconnect(room)
+			room.disconnect()
 
 			return
 		}
@@ -231,12 +91,7 @@ func roomHandler(w http.ResponseWriter, r *http.Request) {
 			}
 
 			log.Println("RoomConfig:", registerRoomMessage.RoomConfig)
-			room.Config = registerRoomMessage.RoomConfig
-
-			result := registerRoom(room)
-			if result {
-				room.State = "ready"
-			}
+			room.register(registerRoomMessage.RoomConfig)
 
 			break
 		case "ready":
@@ -266,153 +121,24 @@ func roomHandler(w http.ResponseWriter, r *http.Request) {
 				continue
 			}
 
-			result, err := handleVerificationCodeResponseMessage(room, verificationCodeResponseMessage)
-			if err != nil {
-				log.Println("Error handling verification code response message:", err)
-				break
-			}
-
-			if result {
-				room.State = "verified"
+			if verificationCodeResponseMessage.Status {
+				room.acceptConnection()
 			} else {
-				// If verification failed, reset to ready state
-				result := registerRoom(room)
-				if result {
-					room.State = "ready"
-				}
+				room.rejectConnection()
 			}
 
 			break
 
 		case "connected":
 			// Room is connected to a plugin
-			// Forward messages between the two connections
 
-			// TODO move this to a separate function
-
-			// Get plugin
-			connManager.Lock()
-			plugin := connManager.plugins[room.PluginID]
-			connManager.Unlock()
-
-			if err := plugin.Conn.WriteMessage(websocket.TextMessage, msg); err != nil {
-				log.Println("write error:", err)
-				handlePluginDisconnect(plugin)
-			}
+			// Forward messages to the plugin
+			plugin := room.getPlugin()
+			plugin.sendMessage(msg)
 
 			break
 		}
 	}
-}
-
-func handleVerificationCodeResponseMessage(room *Room, msg VerificationCodeResponseMessage) (bool, error) {
-	status := msg.Status
-
-	// Get plugin
-	connManager.Lock()
-	plugin := connManager.plugins[room.PluginID]
-	connManager.Unlock()
-
-	if status {
-		// Send VerificationCodeAcceptedMessage to plugin
-		verificationCodeAcceptedMessage := VerificationCodeAcceptedMessage{
-			Type:       MessageTypeVerificationCodeAccepted,
-			RoomConfig: room.Config,
-		}
-		verificationCodeAcceptedMessageJSON, _ := json.Marshal(verificationCodeAcceptedMessage)
-
-		if err := plugin.Conn.WriteMessage(websocket.TextMessage, verificationCodeAcceptedMessageJSON); err != nil {
-			log.Println("write error to plugin:", err)
-			handlePluginDisconnect(plugin)
-			return false, err
-		}
-		log.Println("verification code accepted")
-
-		return true, nil
-	} else {
-		// Send VerificationCodeRejectedMessage to plugin
-		verificationCodeRejectedMessage := VerificationCodeRejectedMessage{
-			Type: MessageTypeVerificationCodeRejected,
-		}
-		verificationCodeRejectedMessageJSON, _ := json.Marshal(verificationCodeRejectedMessage)
-
-		if err := plugin.Conn.WriteMessage(websocket.TextMessage, verificationCodeRejectedMessageJSON); err != nil {
-			log.Println("write error to plugin:", err)
-			handlePluginDisconnect(plugin)
-			return false, err
-		}
-
-		// Reset plugin roomID
-		plugin.RoomID = ""
-
-		return false, nil
-	}
-}
-
-func registerRoom(room *Room) bool {
-	// Generate PIN
-	pin, err := generatePIN()
-	if err != nil {
-		log.Println("Error generating PIN:", err)
-		handleRoomDisconnect(room)
-		return false
-	}
-
-	connManager.Lock()
-	connManager.pinToRoom[pin] = room.Id
-	connManager.Unlock()
-
-	// Create RoomPINMessage
-	roomPINMessageJSON, err := createPairingPINMessage(pin)
-	if err != nil {
-		room.Conn.Close()
-		// Remove room from connection manager
-		connManager.Lock()
-		delete(connManager.rooms, room.Id)
-		delete(connManager.pinToRoom, pin)
-		connManager.Unlock()
-		return false
-	}
-
-	if err := room.Conn.WriteMessage(websocket.TextMessage, roomPINMessageJSON); err != nil {
-		log.Println("write error:", err)
-		room.Conn.Close()
-
-		// Remove room from connection manager
-		connManager.Lock()
-		delete(connManager.rooms, room.Id)
-		delete(connManager.pinToRoom, pin)
-		connManager.Unlock()
-		return false
-	}
-
-	log.Printf("Room connected with pin: %s, roomID: %s", pin, room.Id)
-
-	currentPIN := pin
-
-	ctx, cancel := context.WithCancel(context.Background())
-
-	go func() {
-		defer cancel()
-		ticker := time.NewTicker(pinRotationInterval * time.Second)
-		defer ticker.Stop()
-
-		for {
-			select {
-			case <-ctx.Done():
-				return
-
-			case <-ticker.C:
-				newPIN := rotatePIN(room, currentPIN)
-				currentPIN = newPIN
-			}
-		}
-	}()
-
-	room.StopPinGen = cancel
-
-	return true
-
 }
 
 func pluginHandler(w http.ResponseWriter, r *http.Request) {
@@ -425,23 +151,15 @@ func pluginHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	// Generate plugin ID
-	var pluginID = uuid.New().String()
-
-	plugin := &Plugin{
-		Id:   pluginID,
-		Conn: pluginConn,
-	}
-
-	connManager.Lock()
-	connManager.plugins[pluginID] = plugin
-	connManager.Unlock()
+	// Create a new plugin
+	plugin := newPlugin(pluginConn)
+	connManager.addPlugin(plugin)
 
 	for {
 		_, msg, err := pluginConn.ReadMessage()
 		if err != nil {
 			log.Println("read error:", err)
-			handlePluginDisconnect(plugin)
+			plugin.disconnect()
 			return
 		}
 
@@ -466,9 +184,7 @@ func pluginHandler(w http.ResponseWriter, r *http.Request) {
 		// Depending on the state of the plugin only handle messages with the correct type
 
 		// Find room plugin is connected to
-		connManager.Lock()
-		room := connManager.rooms[plugin.RoomID]
-		connManager.Unlock()
+		room := plugin.getRoom()
 
 		// If not connected to a room
 		if room == nil {
@@ -493,23 +209,12 @@ func pluginHandler(w http.ResponseWriter, r *http.Request) {
 				continue
 			}
 
-			room := handlePairingPINMessage(plugin, pairingPINUserInputMessage)
-			// room is null (invalid pin) wait for another attempt
+			// Try to pair with a room
+			room := plugin.pairWithRoom(pairingPINUserInputMessage.PIN)
+
+			// If room was not found, wait for next attempt
 			if room == nil {
 				continue
-			}
-
-			plugin.RoomID = room.Id
-			room.PluginID = plugin.Id
-			log.Println("Connected to room with ID: ", room.Id)
-
-			success := handleRoomVerification(plugin)
-			if success {
-				// Room is in verifying state
-				room.State = "verifying"
-			} else {
-				// Failed to enter verifying state
-				// TODO: Error handling, reset state
 			}
 
 		} else {
@@ -539,239 +244,20 @@ func pluginHandler(w http.ResponseWriter, r *http.Request) {
 					continue
 				}
 
-				forwardedLinks := handleRoomLinksMessage(plugin, joinURLsMessage)
-
-				if forwardedLinks == false {
-					log.Println("Error forwarding room links")
-					continue
-				}
-
-				log.Println("Successfully forwarded room links")
-
-				// Room is connected to a plugin
-				// Change state to connected
-				room.State = "connected"
+				room.connect(joinURLsMessage.JoinURLs)
 
 				break
 			case "connected":
 				// Room is connected to a plugin
-				// Forward messages between the two connections
 
-				// TODO move this to a separate function
-
-				if err := room.Conn.WriteMessage(websocket.TextMessage, msg); err != nil {
-					log.Println("write error:", err)
-					handleRoomDisconnect(room)
-				}
+				// Forward messages to the room
+				room.sendMessage(msg)
 
 				break
 			}
 
 		}
 	}
-}
-
-func handleRoomLinksMessage(plugin *Plugin, msg JoinURLsMessage) bool {
-	// Get room
-	connManager.Lock()
-	room := connManager.rooms[plugin.RoomID]
-	connManager.Unlock()
-
-	// Forward the JoinURLs message to the room
-	roomConn := room.Conn
-	messageJSON, _ := json.Marshal(msg)
-	if err := roomConn.WriteMessage(websocket.TextMessage, messageJSON); err != nil {
-		log.Println("write error to room:", err)
-		handleRoomDisconnect(room)
-		return false
-	}
-
-	return true
-}
-
-func handlePairingPINMessage(plugin *Plugin, pairingPINUserInputMessage PairingPINUserInputMessage) *Room {
-	pin := pairingPINUserInputMessage.PIN
-
-	connManager.Lock()
-	roomID, exists := connManager.pinToRoom[pin]
-	if !exists {
-		log.Printf("Invalid pin received: %s", pin)
-		connManager.Unlock()
-
-		// Respond with an error message
-		message := PairingPINFailedMessage{
-			Type: MessageTypePairingPINFailed,
-		}
-
-		messageJSON, _ := json.Marshal(message)
-
-		if err := plugin.Conn.WriteMessage(websocket.TextMessage, messageJSON); err != nil {
-			log.Println("write error:", err)
-			plugin.Conn.Close()
-		}
-
-		return nil
-	}
-
-	room := connManager.rooms[roomID]
-
-	// Once paired, remove the pin from mapping
-	delete(connManager.pinToRoom, pin)
-
-	// Stop further PIN rotations
-	connManager.rooms[roomID].StopPinGen()
-	connManager.Unlock()
-
-	log.Printf("Resolved room with roomID: %s for pin: %s", roomID, pin)
-
-	return room
-}
-
-func handleRoomDisconnect(room *Room) {
-	log.Printf("Room disconnected with ID: %s", room.Id)
-
-	// Room disconnected, close the connection
-	room.Conn.Close()
-
-	log.Printf("Connection closed, start cleanup")
-
-	// If room is ready, stop PIN generation
-	if room.State == "ready" {
-		log.Println("Stop pin gen")
-		// Stop PIN generation
-		room.StopPinGen()
-	}
-
-	// If room is registered, remove it from the connection manager
-	if room.State != "new" {
-
-		log.Printf("Remove room from connection manager")
-		connManager.Lock()
-		delete(connManager.rooms, room.Id)
-		connManager.Unlock()
-	}
-
-	log.Println("Check if room was connected to a plugin")
-	// If room is connected to a plugin, notify the plugin
-	connManager.Lock()
-	plugin := connManager.plugins[room.PluginID]
-	connManager.Unlock()
-
-	log.Printf("Plugin ID: %s", room.PluginID)
-
-	if plugin != nil {
-		// Send RoomDisconnected message to plugin
-		roomDisconnectedMessage := RoomDisconnectedMessage{
-			Type: MessageTypeRoomDisconnected,
-		}
-		roomDisconnectedMessageJSON, _ := json.Marshal(roomDisconnectedMessage)
-
-		if err := plugin.Conn.WriteMessage(websocket.TextMessage, roomDisconnectedMessageJSON); err != nil {
-			log.Println("write error to plugin:", err)
-			handlePluginDisconnect(plugin)
-		}
-
-		// Reset plugin roomID
-		plugin.RoomID = ""
-
-	}
-
-	// Remove room from connection manager
-	connManager.Lock()
-	delete(connManager.rooms, room.Id)
-	connManager.Unlock()
-
-	log.Printf("Room removed from connection manager")
-}
-
-func handlePluginDisconnect(plugin *Plugin) {
-	log.Printf("Plugin disconnected with ID: %s", plugin.Id)
-
-	// Plugin disconnected, close the connection
-	plugin.Conn.Close()
-
-	log.Printf("Connection closed, start cleanup")
-
-	// Check if plugin was connected to a room
-	if plugin.RoomID != "" {
-		// Get room
-		connManager.Lock()
-		room := connManager.rooms[plugin.RoomID]
-		connManager.Unlock()
-
-		// If room is connected to the plugin, notify the room
-		if room != nil {
-			// Send PluginDisconnected message to room
-			pluginDisconnectedMessage := PluginDisconnectedMessage{
-				Type: MessageTypePluginDisconnected,
-			}
-			pluginDisconnectedMessageJSON, _ := json.Marshal(pluginDisconnectedMessage)
-
-			if err := room.Conn.WriteMessage(websocket.TextMessage, pluginDisconnectedMessageJSON); err != nil {
-				log.Println("write error to room:", err)
-				handleRoomDisconnect(room)
-			} else {
-				// Reset room pluginID
-				room.PluginID = ""
-
-				// Reset room state to ready
-				result := registerRoom(room)
-				if result {
-					room.State = "ready"
-				}
-			}
-
-		}
-	}
-
-	// Remove plugin from connection manager
-	connManager.Lock()
-	delete(connManager.plugins, plugin.Id)
-	connManager.Unlock()
-}
-
-func handleRoomVerification(plugin *Plugin) bool {
-	// Get room
-	connManager.Lock()
-	room := connManager.rooms[plugin.RoomID]
-	connManager.Unlock()
-
-	roomConn := room.Conn
-
-	// Generate verification code
-	var verificationCode, _ = generateVerificationCode()
-
-	log.Printf("Generated verification code: %s", verificationCode)
-
-	// Generate PairingPINFound message
-	pairingPINFoundMessage := PairingPINFoundMessage{
-		Type:             MessageTypePairingPINFound,
-		VerificationCode: verificationCode,
-	}
-	pairingPINFoundMessageJSON, _ := json.Marshal(pairingPINFoundMessage)
-
-	// Send PairingPINFound message to plugin
-	if err := plugin.Conn.WriteMessage(websocket.TextMessage, pairingPINFoundMessageJSON); err != nil {
-		log.Println("write error to room:", err)
-		handlePluginDisconnect(plugin)
-		return false
-	}
-
-	// Generate VerificationCodeMessage
-	verificationCodeMessage := VerificationCodeMessage{
-		Type:             MessageTypeVerificationCode,
-		VerificationCode: verificationCode,
-	}
-	verificationCodeMessageJSON, _ := json.Marshal(verificationCodeMessage)
-
-	// Send VerificationCodeMessage to room
-	if err := roomConn.WriteMessage(websocket.TextMessage, verificationCodeMessageJSON); err != nil {
-		log.Println("write error to room:", err)
-		handleRoomDisconnect(room)
-		return false
-	}
-
-	return true
 }
 
 func main() {
